pyproject.toml
[tool.poetry]
name = "nonebot-plugin-ocr"
version = "0.1.0"
description = "OCR plugin for nonebot2"
authors = ["NewYearPrism"]
maintainers = ["NewYearPrism"]
readme = "README.md"
packages = [{include = "nonebot_plugin_ocr"}]

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.0.0-beta.5"
nonebot-adapter-onebot = "^2.1.3"
tomli = "^2.0.1"
httpx = "^0.23.0"
cache3 = "^0.3.1"

[tool.poetry.dev-dependencies]
nonebug = "^0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
